+++ src/ghc/Parsley/Internal/Core/Defunc.hs
@@ -37,23 +37,47 @@ data Defunc a where
   -- | Wraps up any value of type `WQ`
   BLACK   :: WQ a -> Defunc a
 
+  -- Syntax constructors
+  {-|
+  Represents the regular Haskell @if@ syntax
+
+  @since 0.1.1.0
+  -}
+  IF_S    :: Defunc Bool -> Defunc a -> Defunc a -> Defunc a
+  {-|
+  Represents a Haskell lambda abstraction
+
+  @since 0.1.1.0
+  -}
+  LAM_S   :: (Defunc a -> Defunc b) -> Defunc (a -> b)
+  {-|
+  Represents a Haskell let binding
+
+  @since 0.1.1.0
+  -}
+  LET_S   :: Defunc a -> (Defunc a -> Defunc b) -> Defunc b
+
 {-|
 This instance is used to manipulate values of `Defunc`.
 
 @since 0.1.0.0
 -}
 instance Quapplicative Defunc where
-  makeQ x qx       = BLACK (makeQ x qx)
-  _val ID          = id
-  _val COMPOSE     = (.)
-  _val FLIP        = flip
-  _val (APP_H f x) = (_val f) (_val x)
-  _val (LIFTED x)  = x
-  _val (EQ_H x)    = ((_val x) ==)
-  _val CONS        = (:)
-  _val CONST       = const
-  _val EMPTY       = []
-  _val (BLACK f)   = _val f
+  makeQ x qx        = BLACK (makeQ x qx)
+  _val ID           = id
+  _val COMPOSE      = (.)
+  _val FLIP         = flip
+  _val (APP_H f x)  = (_val f) (_val x)
+  _val (LIFTED x)   = x
+  _val (EQ_H x)     = ((_val x) ==)
+  _val CONS         = (:)
+  _val CONST        = const
+  _val EMPTY        = []
+  _val (BLACK f)    = _val f
+  -- Syntax
+  _val (IF_S c t e) = if _val c then _val t else _val e
+  _val (LAM_S f)    = \x -> _val (f (makeQ x undefined))
+  _val (LET_S x f)  = let y = _val x in _val (f (makeQ y undefined))
   _code = genDefunc
   (>*<) = APP_H
 
@@ -90,19 +114,24 @@ ap :: Defunc (a -> b) -> Defunc a -> Defunc b
 ap f x = APP_H f x
 
 genDefunc :: Defunc a -> Code a
-genDefunc ID              = [|| \x -> x ||]
-genDefunc COMPOSE         = [|| \f g x -> f (g x) ||]
-genDefunc FLIP            = [|| \f x y -> f y x ||]
-genDefunc (COMPOSE_H f g) = [|| \x -> $$(genDefunc1 (COMPOSE_H f g) [||x||]) ||]
-genDefunc CONST           = [|| \x _ -> x ||]
-genDefunc FLIP_CONST      = [|| \_ y -> y ||]
-genDefunc (FLIP_H f)      = [|| \x -> $$(genDefunc1 (FLIP_H f) [||x||]) ||]
-genDefunc (APP_H f x)     = genDefunc1 f (genDefunc x)
-genDefunc (LIFTED x)      = [|| x ||]
-genDefunc (EQ_H x)        = [|| \y -> $$(genDefunc1 (EQ_H x) [||y||]) ||]
-genDefunc CONS            = [|| \x xs -> x : xs ||]
-genDefunc EMPTY           = [|| [] ||]
-genDefunc (BLACK f)       = _code f
+genDefunc ID                        = [|| \x -> x ||]
+genDefunc COMPOSE                   = [|| \f g x -> f (g x) ||]
+genDefunc FLIP                      = [|| \f x y -> f y x ||]
+genDefunc (COMPOSE_H f g)           = [|| \x -> $$(genDefunc1 (COMPOSE_H f g) [||x||]) ||]
+genDefunc CONST                     = [|| \x _ -> x ||]
+genDefunc FLIP_CONST                = [|| \_ y -> y ||]
+genDefunc (FLIP_H f)                = [|| \x -> $$(genDefunc1 (FLIP_H f) [||x||]) ||]
+genDefunc (APP_H f x)               = genDefunc1 f (genDefunc x)
+genDefunc (LIFTED x)                = [|| x ||]
+genDefunc (EQ_H x)                  = [|| \y -> $$(genDefunc1 (EQ_H x) [||y||]) ||]
+genDefunc CONS                      = [|| \x xs -> x : xs ||]
+genDefunc EMPTY                     = [|| [] ||]
+genDefunc (IF_S (LIFTED True) t _)  = genDefunc t
+genDefunc (IF_S (LIFTED False) _ e) = genDefunc e
+genDefunc (IF_S c t e)              = [|| if $$(genDefunc c) then $$(genDefunc t) else $$(genDefunc e) ||]
+genDefunc (LAM_S f)                 = [|| \x -> $$(genDefunc1 (LAM_S f) [||x||]) ||]
+genDefunc (LET_S x f)               = [|| let y = $$(genDefunc x) in $$(genDefunc (f (makeQ undefined [||y||]))) ||]
+genDefunc (BLACK f)                 = _code f
 
 genDefunc1 :: Defunc (a -> b) -> Code a -> Code b
 genDefunc1 ID              qx = qx
@@ -115,6 +144,7 @@ genDefunc1 CONST           qx = [|| \_ -> $$qx ||]
 genDefunc1 FLIP_CONST      _  = genDefunc ID
 genDefunc1 (FLIP_H f)      qx = [|| \y -> $$(genDefunc2 (FLIP_H f) qx [||y||]) ||]
 genDefunc1 (EQ_H x)        qy = [|| $$(genDefunc x)  == $$qy ||]
+genDefunc1 (LAM_S f)       qx = genDefunc (f (makeQ undefined qx))
 genDefunc1 f               qx = [|| $$(genDefunc f) $$qx ||]
 
 genDefunc2 :: Defunc (a -> b -> c) -> Code a -> Code b -> Code c
@@ -126,7 +156,7 @@ genDefunc2 CONST           qx _   = qx
 genDefunc2 FLIP_CONST      _  qy  = qy
 genDefunc2 (FLIP_H f)      qx qy  = genDefunc2 f qy qx
 genDefunc2 CONS            qx qxs = [|| $$qx : $$qxs ||]
-genDefunc2 f               qx qy  = [|| $$(genDefunc f) $$qx $$qy ||]
+genDefunc2 f               qx qy  = [|| $$(genDefunc1 f qx) $$qy ||]
 
 instance Show (Defunc a) where
   show COMPOSE = "(.)"
@@ -140,4 +170,6 @@ instance Show (Defunc a) where
   show EMPTY = "[]"
   show CONS = "(:)"
   show CONST = "const"
+  show (IF_S c b e) = concat ["(if ", show c, " then ", show b, " else ", show e, ")"]
+  show (LAM_S _) = "f"
   show _ = "x"